version 1.0
Removed all the testing data.
Fixed the fragmentation and added projectileCount to the damage.
Added sortAmmo.
Added a start file where it calls all the other JS files in the right order.
Changed the apiCall name.
Added more calibers to the json file.

FIXED a lot of freaking spelling mistakes.